feat: Enable containerized neovim development and local installation
Introduce a Docker setup for neovim, alongside an installation script
that allows for local configuration.

The Docker setup provides an isolated environment for building and
testing neovim configurations. The installation script then symlinks the
configuration for use in the local environment. This ensures consistency
between development and production and simplifies configuration
management. A .gitignore file is included to ignore temporary files.